- Rename editor to recipe_editor - Fix typo in recipe_description - Add keyboardType to SearchAnchor
