+++ lib/components/recipe_editor/dialogs/d_durations.dart
@@ -1,5 +1,5 @@
 import 'package:flavormate/components/dialogs/t_full_dialog.dart';
-import 'package:flavormate/components/editor/dialogs/d_duration.dart';
+import 'package:flavormate/components/recipe_editor/dialogs/d_duration.dart';
 import 'package:flavormate/components/t_button.dart';
 import 'package:flavormate/components/t_card.dart';
 import 'package:flavormate/components/t_column.dart';
