+++ lib/components/recipe_editor/dialogs/d_instruction_group.dart
@@ -1,6 +1,6 @@
 import 'package:collection/collection.dart';
 import 'package:flavormate/components/dialogs/t_full_dialog.dart';
-import 'package:flavormate/components/editor/dialogs/d_instruction.dart';
+import 'package:flavormate/components/recipe_editor/dialogs/d_instruction.dart';
 import 'package:flavormate/components/t_column.dart';
 import 'package:flavormate/extensions/e_string.dart';
 import 'package:flavormate/l10n/generated/l10n.dart';
