+++ lib/pages/editor/_id.dart
@@ -1,15 +1,15 @@
-import 'package:flavormate/components/editor/dialogs/d_categories.dart';
-import 'package:flavormate/components/editor/dialogs/d_common.dart';
-import 'package:flavormate/components/editor/dialogs/d_course.dart';
-import 'package:flavormate/components/editor/dialogs/d_diet.dart';
-import 'package:flavormate/components/editor/dialogs/d_durations.dart';
-import 'package:flavormate/components/editor/dialogs/d_images.dart';
-import 'package:flavormate/components/editor/dialogs/d_ingredient_groups.dart';
-import 'package:flavormate/components/editor/dialogs/d_instruction_groups.dart';
-import 'package:flavormate/components/editor/dialogs/d_preview.dart';
-import 'package:flavormate/components/editor/dialogs/d_serving.dart';
-import 'package:flavormate/components/editor/dialogs/d_tags.dart';
-import 'package:flavormate/components/editor/t_progress.dart';
+import 'package:flavormate/components/recipe_editor/dialogs/d_categories.dart';
+import 'package:flavormate/components/recipe_editor/dialogs/d_common.dart';
+import 'package:flavormate/components/recipe_editor/dialogs/d_course.dart';
+import 'package:flavormate/components/recipe_editor/dialogs/d_diet.dart';
+import 'package:flavormate/components/recipe_editor/dialogs/d_durations.dart';
+import 'package:flavormate/components/recipe_editor/dialogs/d_images.dart';
+import 'package:flavormate/components/recipe_editor/dialogs/d_ingredient_groups.dart';
+import 'package:flavormate/components/recipe_editor/dialogs/d_instruction_groups.dart';
+import 'package:flavormate/components/recipe_editor/dialogs/d_preview.dart';
+import 'package:flavormate/components/recipe_editor/dialogs/d_serving.dart';
+import 'package:flavormate/components/recipe_editor/dialogs/d_tags.dart';
+import 'package:flavormate/components/recipe_editor/t_progress.dart';
 import 'package:flavormate/components/riverpod/r_struct.dart';
 import 'package:flavormate/components/t_app_bar.dart';
 import 'package:flavormate/components/t_button.dart';
